Divide components for HMR
