+++ libs/sr/ui/src/Provider/CharacterProvider.tsx
@@ -1,0 +1,19 @@
+import type { CharacterKey } from '@genshin-optimizer/sr/consts'
+import type { ReactNode } from 'react'
+import { useMemo, useState } from 'react'
+import type { CharacterContextObj } from '../Context/CharacterContext';
+import { CharacterContext } from '../Context/CharacterContext'
+
+export function CharacterProvider({ children }: { children: ReactNode }) {
+  const [characterKey, setCharacterKey] = useState<CharacterKey | ''>('')
+  const characterContextObj: CharacterContextObj = useMemo(
+    () => ({ characterKey, setCharacterKey }),
+    [characterKey]
+  )
+
+  return (
+    <CharacterContext.Provider value={characterContextObj}>
+      {children}
+    </CharacterContext.Provider>
+  )
+}
