+++ libs/sr/ui/src/index.tsx
@@ -3,6 +3,7 @@ export * from './Context'
 export * from './Hook'
 export * from './LightCone'
 export * from './Optimization'
+export * from './Provider'
 export * from './Relic'
 export * from './Settings'
 export * from './Team'
